+++ .gitignore
@@ -1,0 +1,4 @@
+.expo
+
+node_modules/
+
