+++ package.json
@@ -17,7 +17,11 @@
     "react": "18.2.0",
     "react-native": "0.72.6",
     "react-native-elements": "^3.4.3",
-    "react-navigation": "^5.0.0"
+    "react-native-gesture-handler": "~2.12.0",
+    "react-native-safe-area-context": "4.6.3",
+    "react-native-svg": "13.9.0",
+    "react-navigation": "^5.0.0",
+    "yarn": "^1.22.21"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
